requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.35.0
